+++ resources/views/tailwind/datatable.blade.php
@@ -1,5 +1,15 @@
 <div
-    @if (is_numeric($refresh)) wire:poll.{{ $refresh }}ms @elseif(is_string($refresh)) wire:poll="{{ $refresh }}" @endif
+    @if (is_numeric($refresh))
+        wire:poll.{{ $refresh }}ms
+    @elseif(is_string($refresh))
+        @if ($refresh === '.keep-alive' || $refresh === 'keep-alive')
+            wire:poll.keep-alive
+        @elseif($refresh === '.visible' || $refresh === 'visible')
+            wire:poll.visible
+        @else
+            wire:poll="{{ $refresh }}"
+        @endif
+    @endif
 >
     @include('livewire-tables::tailwind.includes.offline')